requirements.txt
# Core
ib_insync>=0.9.86
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Database
aiosqlite>=0.19.0

# Utilities
pytz>=2023.3
tzdata>=2023.3
nest_asyncio>=1.5.0

# AI
google-generativeai>=0.3.0
anthropic>=0.7.0
apscheduler>=3.10.0
# Data & ML
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# CI/CD
ruff>=0.1.0
mypy>=1.6.0
types-requests
types-pytz
pytest>=7.0.0
pytest-asyncio>=0.21.0
